% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitters.R
\docType{data}
\name{hitters}
\alias{hitters}
\title{KBO Hitters' stats in 2020}
\format{
Tibble with columns
\describe{
\item{선수명}{Player Name}
\item{팀명}{Team Name}
}
}
\source{
\url{http://www.kbreport.com/leader/main}
}
\usage{
hitters
}
\description{
296 Players from 10 teams have 17 stats.
}
\examples{
  hitters
}
\keyword{datasets}
