% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KBO2020-package.R
\docType{package}
\name{KBO2020-package}
\alias{KBO2020}
\alias{KBO2020-package}
\title{KBO2020: KBO Hitters and Pitchers' Stats in 2020}
\description{
KBO Hitters and Pitchers' Stats in 2020
}
\author{
\strong{Maintainer}: DrKevin Na \email{drkevin2022@gmail.com} (\href{https://orcid.org/0000-0003-4132-4784}{ORCID})

}
\keyword{internal}
