% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchers.R
\docType{data}
\name{pitchers}
\alias{pitchers}
\title{KBO Pitchers' stats in 2020}
\format{
Tibble with columns
\describe{
\item{선수명}{Player Name}
\item{팀명}{Team Name}
}
}
\source{
\url{http://www.kbreport.com/leader/main}
}
\usage{
pitchers
}
\description{
283 Players from 10 teams have 18 stats.
}
\examples{
  pitchers
}
\keyword{datasets}
