% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KBO2020-package.R
\docType{package}
\name{KBO2020-package}
\alias{KBO2020}
\alias{KBO2020-package}
\title{KBO2020: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
